+++ packages/core/framework/src/medusa-app-loader.ts
@@ -310,13 +310,25 @@ export class MedusaAppLoader {
    * Load all modules and bootstrap all the modules and links to be ready to be consumed
    * @param config
    */
-  async load(config = { registerInContainer: true }): Promise<MedusaAppOutput> {
+  async load(
+    config: {
+      registerInContainer?: boolean
+      schemaOnly?: boolean
+      migrationOnly?: boolean
+    } = {
+      registerInContainer: true,
+      schemaOnly: false,
+      migrationOnly: false,
+    }
+  ): Promise<MedusaAppOutput> {
     const configModule: ConfigModule = this.#container.resolve(
       ContainerRegistrationKeys.CONFIG_MODULE
     )
 
     const { sharedResourcesConfig, injectedDependencies } =
-      this.prepareSharedResourcesAndDeps()
+      !config.migrationOnly && !config.schemaOnly
+        ? this.prepareSharedResourcesAndDeps()
+        : {}
 
     this.#container.register(
       ContainerRegistrationKeys.REMOTE_QUERY,
@@ -343,6 +355,8 @@ export class MedusaAppLoader {
       injectedDependencies,
       medusaConfigPath: this.#medusaConfigPath,
       cwd: this.#cwd,
+      migrationOnly: config.migrationOnly,
+      schemaOnly: config.schemaOnly,
     })
 
     if (!config.registerInContainer) {