+++ packages/medusa/src/api/admin/translations/middlewares.ts
@@ -2,16 +2,30 @@ import {
   MiddlewareRoute,
   validateAndTransformBody,
   validateAndTransformQuery,
-} from "@medusajs/framework"
+} from "@medusajs/framework/http"
+import { PolicyOperation } from "@medusajs/framework/utils"
+import { DEFAULT_BATCH_ENDPOINTS_SIZE_LIMIT } from "../../../utils"
+import * as QueryConfig from "./query-config"
+import { Entities } from "./query-config"
 import {
   AdminBatchTranslations,
+  AdminBatchTranslationSettings,
   AdminGetTranslationsParams,
+  AdminTranslationEntitiesParams,
+  AdminTranslationSettingsParams,
   AdminTranslationStatistics,
 } from "./validators"
-import * as QueryConfig from "./query-config"
-import { DEFAULT_BATCH_ENDPOINTS_SIZE_LIMIT } from "../../../utils"
 
 export const adminTranslationsRoutesMiddlewares: MiddlewareRoute[] = [
+  {
+    matcher: "/admin/translations/*",
+    policies: [
+      {
+        resource: Entities.translation,
+        operation: PolicyOperation.read,
+      },
+    ],
+  },
   {
     method: ["GET"],
     matcher: "/admin/translations",
@@ -29,6 +43,16 @@ export const adminTranslationsRoutesMiddlewares: MiddlewareRoute[] = [
       sizeLimit: DEFAULT_BATCH_ENDPOINTS_SIZE_LIMIT,
     },
     middlewares: [validateAndTransformBody(AdminBatchTranslations)],
+    policies: [
+      {
+        resource: Entities.translation,
+        operation: PolicyOperation.create,
+      },
+      {
+        resource: Entities.translation,
+        operation: PolicyOperation.update,
+      },
+    ],
   },
   {
     method: ["GET"],
@@ -38,6 +62,33 @@ export const adminTranslationsRoutesMiddlewares: MiddlewareRoute[] = [
   {
     method: ["GET"],
     matcher: "/admin/translations/settings",
-    middlewares: [],
+    middlewares: [
+      validateAndTransformQuery(AdminTranslationSettingsParams, {}),
+    ],
+  },
+  {
+    method: ["POST"],
+    matcher: "/admin/translations/settings/batch",
+    middlewares: [validateAndTransformBody(AdminBatchTranslationSettings)],
+    policies: [
+      {
+        resource: Entities.translation_setting,
+        operation: PolicyOperation.create,
+      },
+      {
+        resource: Entities.translation_setting,
+        operation: PolicyOperation.update,
+      },
+    ],
+  },
+  {
+    method: ["GET"],
+    matcher: "/admin/translations/entities",
+    middlewares: [
+      validateAndTransformQuery(
+        AdminTranslationEntitiesParams,
+        QueryConfig.listTransformQueryConfig
+      ),
+    ],
   },
 ]