+++ packages/medusa/src/api/store/product-variants/route.ts
@@ -4,7 +4,6 @@ import {
 } from "@medusajs/framework/http"
 import { HttpTypes, QueryContextType } from "@medusajs/framework/types"
 import {
-  applyTranslations,
   ContainerRegistrationKeys,
   QueryContext,
 } from "@medusajs/framework/utils"
@@ -12,9 +11,8 @@ import { wrapVariantsWithInventoryQuantityForSalesChannel } from "../../utils/mi
 import { StoreRequestWithContext } from "../types"
 import { wrapVariantsWithTaxPrices } from "./helpers"
 
-type StoreVariantListRequest =
-  StoreRequestWithContext<HttpTypes.StoreProductVariantParams> &
-    AuthenticatedMedusaRequest<HttpTypes.StoreProductVariantParams>
+type StoreVariantListRequest<T = HttpTypes.StoreProductVariantParams> =
+  StoreRequestWithContext<T> & AuthenticatedMedusaRequest<T>
 
 /**
  * @since 2.11.2
@@ -52,14 +50,9 @@ export const GET = async (
       cache: {
         enable: true,
       },
+      locale: req.locale,
     }
   )
-
-  await applyTranslations({
-    localeCode: req.locale,
-    objects: variants,
-    container: req.scope,
-  })
 
   if (withInventoryQuantity) {
     await wrapVariantsWithInventoryQuantityForSalesChannel(req, variants)